# openapi.yaml
openapi: 3.1.0
info:
  title: Retro Casino API
  version: 0.1.0
  description: |
    Play-money arcade casino. The server decides every outcome; the client
    sends a bet and receives a result. All credit values are int64 and are
    never floats anywhere in the system.

    Authentication uses an opaque session token in an httpOnly, SameSite=Lax,
    Secure cookie (`retro_session`). The API is only reachable through the
    Next.js BFF in production; no public CORS.

    WebSocket messages (one connection per player, multiplexed across rooms)
    are documented informally at the bottom of this file and are not part of
    the generated TS client contract.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/guest:
    post:
      operationId: createGuest
      tags: [auth]
      summary: Create a guest session
      description: Creates a guest user with 1000 credits, a wallet, and a server seed. Sets the session cookie.
      responses:
        "200":
          description: Guest created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "429":
          $ref: "#/components/responses/Error"
  /api/v1/auth/keycloak/session:
    post:
      operationId: keycloakSession
      tags: [auth]
      summary: Establish the app identity from a Keycloak access token
      description: >-
        Called by the BFF after an OIDC authorization-code exchange. Verifies
        the access token against the realm JWKS, provisions the local user on
        first login, and upgrades any guest session cookie in place — the
        same user row (wallet, bets, seeds) is retained and the guest
        session is revoked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessToken]
              properties:
                accessToken: { type: string }
      responses:
        "200":
          description: Identity established
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
...
  /api/v1/auth/sessions/{id}:
    delete:
      operationId: revokeSession
      tags: [auth]
      summary: Revoke one session
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/me:
    get:
      operationId: getMe
      tags: [user]
      summary: Current user, wallet balance, and role
      description: >-
        Resolves a Keycloak Bearer token (Authorization header) first, then
        the guest session cookie.
      responses:
        "200":
          description: Current identity
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateMe
      tags: [user]
      summary: Edit the caller's profile (display name, avatar preset)
      description: >-
        Display names are 3-20 chars (letters, digits, spaces, _ and -),
        case-insensitively unique, and cooldown-limited to one change per
        24h (the first rename is exempt). Guests may rename too; the
        Keycloak write-back only applies to registered accounts.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdateRequest" }
      responses:
        "200":
          description: Updated identity and balance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: Rename cooldown active
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: Display name already taken
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/me/avatar:
    put:
      operationId: uploadAvatar
      tags: [user]
      summary: Upload an avatar (registered players only)
      description: >-
        Body is raw PNG or JPEG bytes (max 1MB). The server center-crops,
        downscales to 64x64 pixels, and stores the result; subsequent
        renders use the bumped avatarVersion to cache-bust.
      requestBody:
        required: true
        content:
          image/png:
            schema: { type: string, format: binary }
          image/jpeg:
            schema: { type: string, format: binary }
      responses:
        "200":
          description: New avatar version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AvatarUploadResponse" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: Guests cannot upload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "415": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteAvatar
      tags: [user]
      summary: Clear the caller's avatar (preset and upload)
      responses:
        "204": { description: Cleared }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/me/self-exclude:
    post:
      operationId: selfExclude
      tags: [user]
      summary: Self-exclude from betting
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days: { type: integer, nullable: true }
      responses:
        "200":
          description: New status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  statusUntil: { type: string, format: date-time, nullable: true }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/me/deposit:
    post:
      operationId: deposit
      tags: [user]
      summary: Claim the hourly play-money top-up
      responses:
        "200":
          description: Deposit result
          content:
            application/json:
              schema:
                type: object
                properties:
                  balanceCredits: { type: integer, format: int64 }
                  claimed: { type: boolean }
                  amountCredits: { type: integer, format: int64 }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/users/{id}/avatar:
    get:
      operationId: getUserAvatar
      tags: [user]
      summary: Serve a player's uploaded avatar image
      description: >-
        404 unless the player uploaded an image and has no preset selected.
        Responses are immutable; the version query cache-busts.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
        - name: v
          in: query
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Avatar image
          content:
            image/png:
              schema: { type: string, format: binary }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/users/{id}/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Public profile slice for a player
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicProfile" }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/games/{id}/play:
    post:
      operationId: play
      tags: [games]
      summary: Place a bet and receive the outcome
      description: |
        Server-authoritative. Runs the engine, debits/credits the wallet, and
        records the bet inside one transaction that takes the wallet row
        FOR UPDATE. Returns the outcome, the new balance, and the fairness
        triple used.

        Idempotency: replaying the same idempotencyKey with an identical body
        returns the original result with no second transaction. The same key
        with a differing body returns 409.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlayRequest" }
      responses:
        "200":
          description: Outcome
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "402": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/games/blackjack/deal:
    post:
      operationId: blackjackDeal
      tags: [games]
      summary: Deal a blackjack hand
      description: |
        Server-authoritative, stateful. Shuffles a full deck from the personal
        fair stream (HMAC Fisher-Yates), deals two cards each, debits the
        stake, and records the bet + hand in one transaction. A natural 21
        (player or dealer) resolves and pays inside the same transaction.
        While the hand is active the dealer's hole card is withheld.

        One active hand per user; an abandoned hand (idle > 5 min)
        auto-stands on the next deal. Idempotent like /play.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BlackjackDealRequest" }
      responses:
        "200":
          description: Hand state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HandResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "402": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
...
  /api/v1/hands/{id}/action:
    post:
      operationId: handAction
      tags: [games]
      summary: Hit, stand, or double the active hand
      description: |
        Replays the hand from the fairness triple + action log, applies the
        action, and persists in one transaction. Double debits the extra
        stake; stand / bust / 21 completes the hand: the dealer plays out and
        the payout credits atomically. Retried idempotency keys return the
        current state without drawing again.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, idempotencyKey]
              properties:
                action: { type: string, enum: [hit, stand, double] }
                idempotencyKey: { type: string, maxLength: 64 }
      responses:
        "200":
          description: Hand state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HandResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "402": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
  /api/v1/bets:
    get:
      operationId: listBets
      tags: [games]
      summary: Paginated bet history, own bets only
      description: Ownership is enforced in SQL (`WHERE user_id = $session_user`), never filtered in a handler.
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Bet page
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items: { $ref: "#/components/schemas/Bet" }
                  nextCursor:
                    type: string
                    nullable: true
...
  /api/v1/fair/rotate:
    post:
      operationId: rotateFair
      tags: [fairness]
      summary: Rotate seeds
      description: Reveals the old server seed and issues a new one with its hash published.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientSeed]
              properties:
                clientSeed: { type: string, maxLength: 256 }
      responses:
        "200":
          description: Old seed revealed, new seed active
          content:
            application/json:
              schema:
                type: object
                required: [revealedServerSeed, serverSeedHash, clientSeed, nonce]
                properties:
                  revealedServerSeed: { type: string }
                  serverSeedHash: { type: string }
                  clientSeed: { type: string }
                  nonce: { type: integer, format: int64 }
  /api/v1/ws:
    get:
      operationId: connectSocket
      tags: [lobby]
      summary: Authenticated WebSocket (same session/Bearer auth as HTTP)
      description: |
        Client messages: subscribe_lobby, unsubscribe_lobby, join_room {slug},
        leave_room. Server messages: lobby_summary (1Hz), room_snapshot (on
        join and reconnect), session_revoked, error. place_bet and cash_out
        arrive with the round games.
      responses:
        "101": { description: Switching Protocols }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/themes:
    post:
      operationId: createTheme
      tags: [themes]
      summary: Generate a theme from a prompt
      description: Calls OpenRouter server-side. The response is validated before it is stored or applied; on any failure the current theme remains. Rate limited per user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt: { type: string, maxLength: 200 }
      responses:
        "201":
          description: Generated theme
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Theme" }
        "400": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
    get:
      operationId: listThemes
      tags: [themes]
      summary: User's saved themes
      responses:
        "200":
          description: Themes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Theme" }
...
  /api/v1/rooms/{slug}:
    get:
      operationId: getRoom
      tags: [lobby]
      summary: Room detail and current round snapshot
      description: Standalone fetchable so deep links to /rooms/{slug} work without visiting the lobby.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Room snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RoomSnapshot" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/chat/messages:
    get:
      operationId: getChatHistory
      tags: [social]
      summary: Newest chat lines for dock mounting
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Chat history, oldest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatHistory" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [social]
      summary: Top-20 aggregate plus the caller's own rank
      parameters:
        - name: metric
          in: query
          schema: { type: string, enum: [biggest_win, net_profit, wagered], default: biggest_win }
        - name: window
          in: query
          schema: { type: string, enum: [daily, weekly, all], default: weekly }
      responses:
        "200":
          description: Leaderboard entries (top 20) and the caller's rank when present
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Leaderboard" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List or search users (moderator+)
      parameters:
        - name: query
          in: query
          schema: { type: string }
          description: Substring match on display name or email; empty lists newest first
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 100 }
      responses:
        "200":
          description: User rows
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserList" }
        "401": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{id}/ban:
    post:
      operationId: adminBan
      tags: [admin]
      summary: Ban or unban a user (moderator+)
      description: Writes to audit_log. Banned users can read but cannot bet; their open sockets are closed.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [banned]
              properties:
                banned: { type: boolean }
      responses:
        "204": { description: Applied }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/admin/users/{id}/adjust:
    post:
      operationId: adminAdjust
      tags: [admin]
      summary: Adjust a balance (admin only)
      description: Writes an ordinary ledger transaction with kind admin_adjust and the actor recorded. Nothing writes wallets.balance_credits without a matching transaction row.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountCredits, reason]
              properties:
                amountCredits:
                  type: integer
                  format: int64
                  description: Signed; negative debits, positive credits
                reason: { type: string, minLength: 1 }
      responses:
        "200":
          description: New balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  balanceCredits: { type: integer, format: int64 }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/admin/audit:
    get:
      operationId: adminAudit
      tags: [admin]
      summary: Audit log (admin only)
      parameters:
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Audit entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/AuditEntry" }
                  nextCursor:
                    type: string
                    nullable: true
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: retro_session
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    Me:
      type: object
      required: [user, balanceCredits]
      properties:
        user: { $ref: "#/components/schemas/User" }
        balanceCredits: { type: integer, format: int64 }
    ProfileUpdateRequest:
      type: object
      properties:
        displayName: { type: string }
        avatarPreset:
          type: string
          description: Preset sprite key; empty string clears the preset
    AvatarUploadResponse:
      type: object
      required: [avatarVersion]
      properties:
        avatarVersion: { type: integer, format: int64 }
    PublicProfile:
      type: object
      required: [id, displayName, avatarPreset, avatarVersion, role, createdAt]
      properties:
        id: { type: integer, format: int64 }
        displayName: { type: string }
        avatarPreset: { type: string }
        avatarVersion: { type: integer, format: int64 }
        role: { type: string }
        createdAt: { type: string, format: date-time }
...
    PlayRequest:
      type: object
      required: [betCredits, clientSeed, idempotencyKey]
      properties:
        betCredits:
          type: integer
          format: int64
          description: One of the server-enforced bet steps (5/10/25/50/100 for slots)
        clientSeed: { type: string, maxLength: 256 }
        idempotencyKey: { type: string, maxLength: 64 }
    PlayResponse:
      type: object
      required: [betId, payoutCredits, balanceCredits, outcome, fairness]
      properties:
        betId: { type: integer, format: int64 }
        payoutCredits: { type: integer, format: int64 }
        balanceCredits: { type: integer, format: int64 }
        outcome:
          description: Game-specific payload rendered by the client (3x3 grid + winning lines for slots)
        fairness: { $ref: "#/components/schemas/FairnessTriple" }
...
    BlackjackDealRequest:
      type: object
      required: [betCredits, idempotencyKey]
      properties:
        betCredits:
          type: integer
          format: int64
          description: One of the game's bet steps
        clientSeed: { type: string, maxLength: 256 }
        idempotencyKey: { type: string, maxLength: 64 }
...
    HandResponse:
      type: object
      required: [hand, balanceCredits, fairness, replay]
      properties:
        hand: { $ref: "#/components/schemas/BlackjackHand" }
        balanceCredits: { type: integer, format: int64 }
        fairness: { $ref: "#/components/schemas/FairnessTriple" }
        replay: { type: boolean }
...
    Bet:
      type: object
      required: [id, gameId, betCredits, payoutCredits, nonce, clientSeed, serverSeedHash, createdAt]
      properties:
        id: { type: integer, format: int64 }
        gameId: { type: string }
        roundId: { type: integer, format: int64 }
        betCredits: { type: integer, format: int64 }
        payoutCredits: { type: integer, format: int64 }
        nonce: { type: integer, format: int64 }
        clientSeed: { type: string }
        serverSeedHash: { type: string }
        outcome: { nullable: true }
        createdAt: { type: string, format: date-time }
    Theme:
      type: object
      required: [id, name, palette, sprites, createdAt]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        palette:
          type: array
          description: Exactly 8 hex colors; index 0 is transparent
          items: { type: string }
        sprites:
          type: array
          description: Exactly 8 sprites, each 16 rows of 16 hex-digit characters
          items:
            type: object
            required: [name, rows]
            properties:
              name: { type: string }
              rows:
                type: array
                items: { type: string }
        createdAt: { type: string, format: date-time }
...